refactor: split read-only vector storage modules
