+++ lib/segment/src/vector_storage/dense/read_only/read_ops.rs
@@ -1,0 +1,70 @@
+use common::bitvec::BitSlice;
+use common::generic_consts::AccessPattern;
+use common::types::PointOffsetType;
+use common::universal_io::UniversalRead;
+
+use super::ReadOnlyChunkedDenseVectorStorage;
+use crate::data_types::named_vectors::CowVector;
+use crate::data_types::primitive::PrimitiveVectorElement;
+use crate::types::{Distance, VectorStorageDatatype};
+use crate::vector_storage::{VectorOffsetType, VectorStorageRead};
+
+impl<T: PrimitiveVectorElement, S: UniversalRead> VectorStorageRead
+    for ReadOnlyChunkedDenseVectorStorage<T, S>
+{
+    fn distance(&self) -> Distance {
+        self.distance
+    }
+
+    fn datatype(&self) -> VectorStorageDatatype {
+        T::datatype()
+    }
+
+    fn is_on_disk(&self) -> bool {
+        self.vectors.is_on_disk()
+    }
+
+    fn total_vector_count(&self) -> usize {
+        self.vectors.len()
+    }
+
+    fn get_vector<P: AccessPattern>(&self, key: PointOffsetType) -> CowVector<'_> {
+        self.vectors
+            .get::<P>(key as VectorOffsetType)
+            .map(|slice| CowVector::from(T::slice_to_float_cow(slice)))
+            .expect("Vector not found")
+    }
+
+    fn read_vectors<P: AccessPattern, U: Copy>(
+        &self,
+        keys: impl IntoIterator<Item = (U, PointOffsetType)>,
+        mut callback: impl FnMut(U, PointOffsetType, CowVector<'_>),
+    ) {
+        let keys = keys
+            .into_iter()
+            .map(|(user_data, point_offset)| ((user_data, point_offset), point_offset, 1));
+
+        for ((user_data, point_offset), vector) in self.vectors.iter_vectors::<P, _>(keys) {
+            let vector = CowVector::from(T::slice_to_float_cow(vector));
+            callback(user_data, point_offset, vector);
+        }
+    }
+
+    fn get_vector_opt<P: AccessPattern>(&self, key: PointOffsetType) -> Option<CowVector<'_>> {
+        self.vectors
+            .get::<P>(key as VectorOffsetType)
+            .map(|slice| CowVector::from(T::slice_to_float_cow(slice)))
+    }
+
+    fn is_deleted_vector(&self, key: PointOffsetType) -> bool {
+        self.deleted.get(key)
+    }
+
+    fn deleted_vector_count(&self) -> usize {
+        self.deleted.count()
+    }
+
+    fn deleted_vector_bitslice(&self) -> &BitSlice {
+        self.deleted.as_bitslice()
+    }
+}
