+++ lib/segment/src/vector_storage/multi_dense/read_only/read_ops.rs
@@ -1,0 +1,75 @@
+use common::bitvec::BitSlice;
+use common::generic_consts::AccessPattern;
+use common::types::PointOffsetType;
+use common::universal_io::UniversalRead;
+
+use super::ReadOnlyChunkedMultiDenseVectorStorage;
+use crate::data_types::named_vectors::CowVector;
+use crate::data_types::primitive::PrimitiveVectorElement;
+use crate::types::{Distance, VectorStorageDatatype};
+use crate::vector_storage::VectorStorageRead;
+use crate::vector_storage::multi_dense::appendable_mmap_multi_dense_vector_storage::{
+    flattened_to_multi_vector, read_multi_vector,
+};
+
+impl<T: PrimitiveVectorElement, S: UniversalRead> VectorStorageRead
+    for ReadOnlyChunkedMultiDenseVectorStorage<T, S>
+{
+    fn distance(&self) -> Distance {
+        self.distance
+    }
+
+    fn datatype(&self) -> VectorStorageDatatype {
+        T::datatype()
+    }
+
+    fn is_on_disk(&self) -> bool {
+        self.vectors.is_on_disk()
+    }
+
+    fn total_vector_count(&self) -> usize {
+        self.offsets.len()
+    }
+
+    fn get_vector<P: AccessPattern>(&self, key: PointOffsetType) -> CowVector<'_> {
+        self.get_vector_opt::<P>(key).expect("vector not found")
+    }
+
+    fn read_vectors<P: AccessPattern, U: Copy>(
+        &self,
+        keys: impl IntoIterator<Item = (U, PointOffsetType)>,
+        mut callback: impl FnMut(U, PointOffsetType, CowVector<'_>),
+    ) {
+        let point_offsets = keys
+            .into_iter()
+            .map(|(user_data, point_offset)| ((user_data, point_offset), point_offset));
+
+        let vectors =
+            super::iter_vectors::<P, _, _, _>(&self.offsets, &self.vectors, point_offsets);
+
+        for ((user_data, point_offset), flattened) in vectors {
+            let vector = CowVector::MultiDense(T::into_float_multivector(
+                flattened_to_multi_vector(flattened, self.vectors.dim()),
+            ));
+
+            callback(user_data, point_offset, vector);
+        }
+    }
+
+    fn get_vector_opt<P: AccessPattern>(&self, key: PointOffsetType) -> Option<CowVector<'_>> {
+        read_multi_vector::<T, P, _>(&self.offsets, &self.vectors, key)
+            .map(|multi| CowVector::MultiDense(T::into_float_multivector(multi)))
+    }
+
+    fn is_deleted_vector(&self, key: PointOffsetType) -> bool {
+        self.deleted.get(key)
+    }
+
+    fn deleted_vector_count(&self) -> usize {
+        self.deleted.count()
+    }
+
+    fn deleted_vector_bitslice(&self) -> &BitSlice {
+        self.deleted.as_bitslice()
+    }
+}
