+++ lib/segment/src/vector_storage/quantized/quantized_vectors/read_only/lifecycle.rs
