+++ lib/segment/src/vector_storage/quantized/quantized_vectors/read_only/mod.rs
@@ -1,4 +1,4 @@
-mod load;
+mod lifecycle;
 mod storage;
 #[cfg(test)]
 mod tests;
